feat: add integration node
