+++ gno.land/pkg/gnoland/app.go
@@ -147,10 +147,9 @@ func InitChainer(baseApp *sdk.BaseApp, acctKpr auth.AccountKeeperI, bankKpr bank
 		genState := req.AppState.(GnoGenesisState)
 		// Parse and set genesis state balances.
 		for _, bal := range genState.Balances {
-			addr, coins := parseBalance(bal)
-			acc := acctKpr.NewAccountWithAddress(ctx, addr)
+			acc := acctKpr.NewAccountWithAddress(ctx, bal.Address)
 			acctKpr.SetAccount(ctx, acc)
-			err := bankKpr.SetCoins(ctx, addr, coins)
+			err := bankKpr.SetCoins(ctx, bal.Address, bal.Value)
 			if err != nil {
 				panic(err)
 			}
