+++ gno.land/pkg/gnoland/genesis.go
@@ -1,0 +1,79 @@
+package gnoland
+
+import (
+	"errors"
+	"fmt"
+	"strings"
+
+	"github.com/gnolang/gno/tm2/pkg/amino"
+	"github.com/gnolang/gno/tm2/pkg/crypto"
+	osm "github.com/gnolang/gno/tm2/pkg/os"
+	"github.com/gnolang/gno/tm2/pkg/std"
+)
+
+func LoadGenesisBalancesFile(path string) ([]Balance, error) {
+	// each balance is in the form: g1xxxxxxxxxxxxxxxx=100000ugnot
+	content := osm.MustReadFile(path)
+	lines := strings.Split(string(content), "\n")
+
+	balances := make([]Balance, 0, len(lines))
+	for _, line := range lines {
+		line = strings.TrimSpace(line)
+
+		// remove comments.
+		line = strings.Split(line, "#")[0]
+		line = strings.TrimSpace(line)
+
+		// skip empty lines.
+		if line == "" {
+			continue
+		}
+
+		parts := strings.Split(line, "=") // <address>=<coin>
+		if len(parts) != 2 {
+			return nil, errors.New("invalid genesis_balance line: " + line)
+		}
+
+		addr, err := crypto.AddressFromBech32(parts[0])
+		if err != nil {
+			return nil, fmt.Errorf("invalid balance addr %s: %w", parts[0], err)
+		}
+
+		coins, err := std.ParseCoins(parts[1])
+		if err != nil {
+			return nil, fmt.Errorf("invalid balance coins %s: %w", parts[1], err)
+		}
+
+		balances = append(balances, Balance{
+			Address: addr,
+			Value:   coins,
+		})
+	}
+
+	return balances, nil
+}
+
+// XXX: we can do something better here
+func LoadGenesisTxsFile(path string, chainID string, genesisRemote string) ([]std.Tx, error) {
+	txs := []std.Tx{}
+	txsBz := osm.MustReadFile(path)
+	txsLines := strings.Split(string(txsBz), "\n")
+	for _, txLine := range txsLines {
+		if txLine == "" {
+			continue // Skip empty line.
+		}
+
+		// Patch the TX.
+		txLine = strings.ReplaceAll(txLine, "%%CHAINID%%", chainID)
+		txLine = strings.ReplaceAll(txLine, "%%REMOTE%%", genesisRemote)
+
+		var tx std.Tx
+		if err := amino.UnmarshalJSON([]byte(txLine), &tx); err != nil {
+			return nil, fmt.Errorf("unable to Unmarshall txs file: %w", err)
+		}
+
+		txs = append(txs, tx)
+	}
+
+	return txs, nil
+}
