+++ gno.land/pkg/integration/node_inmemory.go
@@ -1,0 +1,5 @@
+package integration
+
+import "github.com/gnolang/gno/tm2/pkg/log"
+
+}
