+++ gno.land/pkg/integration/testing_integration.go
@@ -13,13 +13,17 @@ import (
 	"time"
 
 	"github.com/gnolang/gno/gno.land/pkg/gnoland"
+	abci "github.com/gnolang/gno/tm2/pkg/bft/abci/types"
+	"github.com/gnolang/gno/tm2/pkg/bft/config"
 	"github.com/gnolang/gno/tm2/pkg/bft/node"
 	"github.com/gnolang/gno/tm2/pkg/bft/types"
 	"github.com/gnolang/gno/tm2/pkg/commands"
+	"github.com/gnolang/gno/tm2/pkg/crypto"
 	"github.com/gnolang/gno/tm2/pkg/crypto/keys"
 	"github.com/gnolang/gno/tm2/pkg/crypto/keys/client"
 	"github.com/gnolang/gno/tm2/pkg/events"
 	"github.com/gnolang/gno/tm2/pkg/log"
+	"github.com/gnolang/gno/tm2/pkg/std"
 	"github.com/rogpeppe/go-internal/testscript"
 )
 
@@ -51,14 +55,14 @@ func SetupGnolandTestScript(t *testing.T, txtarDir string) testscript.Params {
 
 	// `gnoRootDir` should point to the local location of the gno repository.
 	// It serves as the gno equivalent of GOROOT.
-	gnoRootDir := gnoland.GuessGnoRootDir()
+	gnoRootDir := gnoland.MustGuessGnoRootDir()
 
 	// `gnoHomeDir` should be the local directory where gnokey stores keys.
 	gnoHomeDir := filepath.Join(tmpdir, "gno")
 
 	// `gnoDataDir` should refer to the local location where the gnoland node
 	// stores its configuration and data.
-	gnoDataDir := filepath.Join(tmpdir, "data")
+	// gnoDataDir := filepath.Join(tmpdir, "data")
 
 	// Testscripts run concurrently by default, so we need to be prepared for that.
 	var muNodes sync.Mutex
@@ -115,9 +119,8 @@ func SetupGnolandTestScript(t *testing.T, txtarDir string) testscript.Params {
 						logger = getTestingLogger(ts, logname)
 					}
 
-					dataDir := filepath.Join(gnoDataDir, sid)
 					var node *node.Node
-					if node, err = execTestingGnoland(t, logger, dataDir, gnoRootDir, args); err == nil {
+					if node, err = execInMemoryGnoland(logger, gnoRootDir); err == nil {
 						nodes[sid] = &testNode{
 							Node:   node,
 							logger: logger,
@@ -139,7 +142,7 @@ func SetupGnolandTestScript(t *testing.T, txtarDir string) testscript.Params {
 
 						// Add default environements.
 						ts.Setenv("RPC_ADDR", laddr)
-						ts.Setenv("GNODATA", gnoDataDir)
+						// ts.Setenv("GNODATA", gnoDataDir)
 
 						const listenerID = "testing_listener"
 
@@ -178,7 +181,7 @@ func SetupGnolandTestScript(t *testing.T, txtarDir string) testscript.Params {
 
 						// Unset gnoland environements.
 						ts.Setenv("RPC_ADDR", "")
-						ts.Setenv("GNODATA", "")
+						// ts.Setenv("GNODATA", "")
 						fmt.Fprintln(ts.Stdout(), "node stopped successfully")
 					}
 				default:
@@ -279,4 +282,70 @@ func getTestingLogger(ts *testscript.TestScript, logname string) log.Logger {
 
 	ts.Logf("starting logger: %q", path)
 	return logger
+}
+
+func execInMemoryGnoland(logger log.Logger, gnoroot string) (*node.Node, error) {
+	txsFile := filepath.Join(gnoroot, "gno.land", "genesis", "genesis_txs.txt")
+	balanceFile := filepath.Join(gnoroot, "gno.land", "genesis", "genesis_balances.txt")
+	exampleDir := filepath.Join(gnoroot, "examples")
+
+	genesisBalances, err := gnoland.LoadGenesisBalancesFile(balanceFile)
+	if err != nil {
+		return nil, fmt.Errorf("unableo to laod genesis balance file: %w", err)
+	}
+
+	bftconfig := config.TestConfig().SetRootDir(gnoroot)
+	bftconfig.Consensus.CreateEmptyBlocks = true
+	bftconfig.Consensus.CreateEmptyBlocksInterval = time.Duration(0)
+	bftconfig.RPC.ListenAddress = "tcp://127.0.0.1:0"
+	bftconfig.P2P.ListenAddress = "tcp://127.0.0.1:0"
+
+	// NOTE: we dont care about giving a correct address here, as it's only visual
+	// XXX: do we care loading this file ?
+	genesisTXs, err := gnoland.LoadGenesisTxsFile(txsFile, bftconfig.ChainID(), bftconfig.RPC.ListenAddress)
+	if err != nil {
+		return nil, fmt.Errorf("unableo to laod genesis balance file: %w", err)
+	}
+
+	// Load example packages
+	pkgs := PackagePath{
+		Creator: crypto.MustAddressFromString(test1Addr),
+		Fee:     std.NewFee(50000, std.MustParseCoin("1000000ugnot")),
+		Path:    exampleDir,
+	}
+
+	config := NodeConfig{
+		Balances:   genesisBalances,
+		GenesisTXs: genesisTXs,
+		BFTConfig:  bftconfig,
+		Packages:   []PackagePath{pkgs},
+	}
+
+	// XXX: make this configurable
+	config.ConsensusParams.Block = &abci.BlockParams{
+		MaxTxBytes:   1000000,  // 1MB,
+		MaxDataBytes: 2000000,  // 2MB,
+		MaxGas:       10000000, // 10M gas
+		TimeIotaMS:   100,      // 100ms
+	}
+
+	node, err := NewNode(logger, config)
+	if err != nil {
+		return nil, fmt.Errorf("unable to start the node: %w", err)
+	}
+
+	return node, node.Start()
+}
+
+func tsValidateError(ts *testscript.TestScript, cmd string, neg bool, err error) {
+	if err != nil {
+		fmt.Fprintf(ts.Stderr(), "%q error: %v\n", cmd, err)
+		if !neg {
+			ts.Fatalf("unexpected %q command failure: %s", cmd, err)
+		}
+	} else {
+		if neg {
+			ts.Fatalf("unexpected %s command success", cmd)
+		}
+	}
 }
